Adding audio and image/video files needed for loading the Gradio UI, and update the UI Python function (#1034)
