+++ AvatarChatbot/tests/test_compose_on_xeon.sh
